api/requirements.txt
Babel==2.6.0
Django==2.0.6
django-filter==1.1.0
django-phonenumber-field==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Markdown==2.6.11
phonenumberslite==8.9.8
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.4
six==1.11.0
